requirements.txt
asyncio==3.4.3
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
protobuf==6.31.1
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
websockets==15.0.1
